sensors/pyproject.toml
[project]
name = "iot-sensors"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "python-dotenv>=1.0.0",
]
